Feat: Add Pickle Level & Add Confirmation Window
Added a progress bar that shows your current pickle level, which progresses every pickle click. Will soon be utilized to get achievements shown on Steam. In addition, a confirmation window when exiting the game will pop up and based on the user can accept or decline. Warning text will appear above the buttons and will vary depending on the action bound to be taken.